requirements.txt
streamlit==1.37.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
pymupdf==1.24.9
markdown==3.6
beautifulsoup4==4.12.3
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
